Use Asynchronous LCD
The original LiquidCrystal library is synchronous, that is, it block everything until the text is sent to the LCD.

This blocks the I2C bus and slows down interruptions.

A new LCD library was developed that queues all the commands and executes then asynchronously.
Also, is the internal queue is full, writeText/writeCommand return it and the caller is expected to retry